sửa lại chổ occupier comment
