+++ Login.h
@@ -7,4 +7,5 @@ void find_suitable_house(Member *currentUser,Global *program);
 void send_request(Member *currenUser, Member *chosenUser);
 void rate_house(Member *chosenUser);
 void showHouse(Global *program);
+void accept_request(Member *currentMember, Global *program);
 #endif
